MAL API First commit
I've already added some stuff, for now it's just an Anime model with a
parser that populates the anime model.

It can get an ID from an anime (from MAL of course.) and parse some info
like Genres. All this is, is a crawler and it works pretty well. I'm not
so sure how fast it is though. I'll have to test it out.